Add option to define a gwt.version in the BuildConfig.
This will trigger the script to download the appropriate jars via the configured Ivy settings.

They will then be copied into a temp directory which is then used as the GWT home.

An optimisation would have them being used in place in the Ivy cache, bypassing the need for a GWT home.

At minimum, this will require the user to set the GWT version and then uncomment the mavenCentral repo.